move AppShell code back into App2.tsx (problem with navbar Accordion not remembering state), change name SwitchToggle -> ColourSchemeToggle
