+++ src/main/webui/src/App2.tsx
@@ -2,40 +2,64 @@ import {
     createContext,
     useContext,
     ReactElement,
-    Context, StrictMode, SyntheticEvent
+    SyntheticEvent,
+    Context, StrictMode
 } from 'react';
 import {
     QueryClient,
     QueryClientProvider,
 } from '@tanstack/react-query';
-import { Person} from "./generated/proposalToolSchemas";
-import TitlePanel from './ProposalEditorView/proposal/Title';
-import OverviewPanel from "./ProposalEditorView/proposal/Overview";
-import NewProposalPanel from './ProposalEditorView/proposal/New';
-import SummaryPanel from "./ProposalEditorView/proposal/Summary";
-import InvestigatorsPanel from "./ProposalEditorView/Investigators/List";
-import AddInvestigatorPanel from "./ProposalEditorView/Investigators/New";
+import { Person} from "./generated/proposalToolSchemas.ts";
+import TitlePanel from './ProposalEditorView/proposal/Title.tsx';
+import OverviewPanel from "./ProposalEditorView/proposal/Overview.tsx";
+import NewProposalPanel from './ProposalEditorView/proposal/New.tsx';
+import SummaryPanel from "./ProposalEditorView/proposal/Summary.tsx";
+import InvestigatorsPanel from "./ProposalEditorView/Investigators/List.tsx";
+import AddInvestigatorPanel from "./ProposalEditorView/Investigators/New.tsx";
 import {
     createBrowserRouter,
-    RouterProvider, useNavigate,
+    Outlet,
+    RouterProvider,
+    useNavigate
 } from 'react-router-dom';
+import { useHistoryState } from "./useHistoryState";
 import TechnicalGoalsPanel from "./ProposalEditorView/technicalGoals/technicalGoalsPanel.tsx";
 import { TargetPanel } from "./ProposalEditorView/targets/targetPanel.tsx";
 import ObservationsPanel from "./ProposalEditorView/observations/observationPanel.tsx";
-import DocumentsPanel from "./ProposalEditorView/proposal/Documents";
-import SubmitPanel from "./ProposalEditorView/proposal/Submit";
+import DocumentsPanel from "./ProposalEditorView/proposal/Documents.tsx";
+import SubmitPanel from "./ProposalEditorView/proposal/Submit.tsx";
 
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import {AuthProvider} from "./auth/Auth.tsx";
 import {
-    Text, Container, Group, ActionIcon
+    AppShell,
+    Text,
+    TextInput,
+    Grid,
+    Burger,
+    ScrollArea,
+    Group,
+    ActionIcon,
+    Tooltip, useMantineTheme, useMantineColorScheme, FileButton, Container
 } from '@mantine/core';
-import AdminPanel from "./admin/adminPanel.tsx";
-import JustificationsPanel from "./ProposalEditorView/justifications/JustificationsPanel.tsx";
-import EditorAppShell from "./EditorAppShell.tsx";
-import ManagementAppShell from "./ManagementAppShell.tsx";
-import * as React from "react";
-import {IconEdit, IconUniverse} from "@tabler/icons-react";
+import {ColourSchemeToggle} from "./ColourSchemeToggle";
+import {
+    IconLicense,
+    IconLogout, IconUniverse
+} from '@tabler/icons-react';
+import {useDisclosure} from "@mantine/hooks";
+import AddButton from './commonButtons/add';
+import DatabaseSearchButton from './commonButtons/databaseSearch';
+import {
+    APP_HEADER_HEIGHT, CLOSE_DELAY, ICON_SIZE,
+    NAV_BAR_DEFAULT_WIDTH, NAV_BAR_LARGE_WIDTH,
+    NAV_BAR_MEDIUM_WIDTH, OPEN_DELAY,
+} from './constants';
+import { handleUploadZip } from './ProposalEditorView/proposal/UploadProposal';
+import UploadButton from './commonButtons/upload';
+import AdminPanel from "./admin/adminPanel";
+import JustificationsPanel from "./ProposalEditorView/justifications/JustificationsPanel";
+import {ProposalList} from "./ProposalList";
 
 /**
  * defines the user context type.
@@ -60,7 +84,7 @@ export type ProposalContextType = {
  * @type {React.Context<UserContextType & ProposalContextType>} the context.
  */
 export const ProposalContext:
-        Context<UserContextType & ProposalContextType> =
+    Context<UserContextType & ProposalContextType> =
     createContext<UserContextType & ProposalContextType>({
         user: {},
         getToken: ()=>{return ""},
@@ -77,18 +101,6 @@ export const useToken = (): string => {
     return useContext(ProposalContext).getToken();
 };
 
-export type CycleContextType = {
-    selectedCycleCode: number;
-}
-export const CycleContext: Context<CycleContextType>  = createContext({
-    selectedCycleCode: 0
-});
-
-export type PSTAppShellProps = {
-    setEditorMode: React.Dispatch<React.SetStateAction<boolean>>
-}
-
-
 /**
  * generates the html for the main app.
  * @return {ReactElement} dynamic html for the main app.
@@ -97,24 +109,28 @@ export type PSTAppShellProps = {
 function App2(): ReactElement {
 
     // get database query client.
-    const queryClient = new QueryClient();
+    const queryClient = new QueryClient()
+
+    // the colour gray used by the tools.
+    const theme = useMantineTheme();
+    const {colorScheme} = useMantineColorScheme();
+
+    const GRAY = theme.colors.gray[6];
 
     // the paths to route to.
-    const router = createBrowserRouter([
+    const router = createBrowserRouter(
+        [
             {
-              path: "/", element: <PSTRoot />,
+                path: "manager",
+                element: <PSTManager />
             },
-            {
-                path: "/editor", element: <PSTEditor />,
+            {path: "/", element: <PSTRoot/>,
                 children: [
                     {index: true, element: <PSTStart/>} ,
                     {
                         path: "admin",
                         element: <AdminPanel />
                     },
-
-                    // ---- Proposal Editor routes -------
-
                     {
                         path: "proposal/new",
                         element: <NewProposalPanel />
@@ -165,73 +181,180 @@ function App2(): ReactElement {
                         element:<SubmitPanel />
                     },
 
-                    // ---- Proposal Management routes -------
-                ]
-            },
-            {
-                path: "manager", element: <PSTManager />,
-            }
-            ],
-        {basename: "/pst/gui/tool/"}
+                ]}], {
+            basename: "/pst/gui/tool/"
+        }
+
     )
 
     return (
         <AuthProvider>
             <StrictMode>
-            <QueryClientProvider client={queryClient}>
-                <RouterProvider router={router}/>
-                <ReactQueryDevtools initialIsOpen={false} />
-            </QueryClientProvider>
+                <QueryClientProvider client={queryClient}>
+                    <RouterProvider router={router}/>
+                    <ReactQueryDevtools initialIsOpen={false} />
+                </QueryClientProvider>
             </StrictMode>
         </AuthProvider>
     );
-
-
-    function PSTRoot() : ReactElement {
-        const navigate = useNavigate();
-        return (
-            <Container fluid>
-                <Group>
-                    <ActionIcon
-                        onClick={(e : SyntheticEvent)=>{e.preventDefault(); navigate("editor")}}
-                    >
-                        <IconEdit/>
-                    </ActionIcon>
-                    <ActionIcon
-                        onClick={(e: SyntheticEvent)=>{e.preventDefault(); navigate("manager")}}
-                    >
-                        <IconUniverse/>
-                    </ActionIcon>
-                </Group>
-            </Container>
-        )
-    }
-
 
     /**
      * main HTML for the UI.
      * @return {ReactElement} the dynamic html for the main UI.
      * @constructor
      */
-    function PSTEditor(): ReactElement {
+    function PSTRoot(): ReactElement {
         const proposalContext = useContext(ProposalContext);
+        const [opened, {toggle}] = useDisclosure();
+        const navigate = useNavigate();
+        // acquire the state setters for proposal title and investigator name.
+        const [proposalTitleFilter, setProposalTitleFilter] = useHistoryState(
+            "proposalTitle", "");
+        const [investigatorNameFilter, setInvestigatorNameFilter] = useHistoryState(
+            "investigatorName", "");
+
+        //active state for the NavLink sections
+
+        /**
+         * resolves the routing for when making a new proposal.
+         *
+         * @param {React.SyntheticEvent} event the event.
+         */
+        function handleAddNew(event: SyntheticEvent): void {
+            event.preventDefault();
+            navigate("proposal/new");
+        }
+
+        /**
+         * resolves the routing for when searching for a proposal.
+         *
+         * @param {React.SyntheticEvent} event the event.
+         */
+        function handleSearch(event: SyntheticEvent): void {
+            event.preventDefault();
+            navigate("/");
+        }
+
         return (
             <ProposalContext.Provider value={proposalContext}>
-                <EditorAppShell />
+                <AppShell
+                    header={{height: APP_HEADER_HEIGHT}}
+                    navbar={{
+                        width: {
+                            base: NAV_BAR_DEFAULT_WIDTH,
+                            md: NAV_BAR_MEDIUM_WIDTH,
+                            lg: NAV_BAR_LARGE_WIDTH},
+                        breakpoint: 'sm',
+                        collapsed: {mobile: !opened},
+                    }}
+                >
+                    <AppShell.Header p="md">
+                        <Grid columns={2}>
+                            <Grid.Col span={1}>
+                                <Group h="100%" px="md" wrap={"nowrap"}>
+                                    <Burger
+                                        opened={opened}
+                                        onClick={toggle}
+                                        hiddenFrom={"sm"}
+                                        size="lg"
+                                        color={GRAY}
+                                        mr="xl"
+                                    />
+                                    <img src={"/pst/gui/polaris4.png"}
+                                         alt="Polaris"
+                                         width={60}/>
+                                    <ActionIcon
+                                        color={"pink"}
+                                        variant={"subtle"}
+                                        onClick={(e: SyntheticEvent)=>{e.preventDefault(); navigate("/manager")}}
+                                    >
+                                        <IconUniverse />
+                                    </ActionIcon>
+                                    <DatabaseSearchButton
+                                        toolTipLabel={
+                                            "Locate proposals by " +
+                                            proposalContext.user.fullName + "."}
+                                        label={"Proposals for " + proposalContext.user.fullName}
+                                        onClickEvent={handleSearch}
+                                    />
+                                    <AddButton toolTipLabel={"new proposal"}
+                                               label={"Create a new proposal"}
+                                               onClickEvent={handleAddNew}/>
+                                    <FileButton onChange={handleUploadZip}
+                                                accept={".zip"}>
+                                        {(props) => <UploadButton
+                                            toolTipLabel="select a file from disk to upload"
+                                            label={"Import"}
+                                            onClick={props.onClick}/>}
+                                    </FileButton>
+                                </Group>
+                            </Grid.Col>
+                            <Grid.Col span={1}>
+                                <Group justify={"flex-end"}>
+                                    {ColourSchemeToggle()}
+                                    <Tooltip label={"logout"}
+                                             openDelay={OPEN_DELAY}
+                                             closeDelay={CLOSE_DELAY}
+                                    >
+                                        <ActionIcon color={"orange.8"}
+                                                    variant={"subtle"}
+                                                    component={"a"}
+                                                    href={"/pst/gui/logout"}
+                                        >
+                                            <IconLogout size={ICON_SIZE}/>
+                                        </ActionIcon>
+                                    </Tooltip>
+
+
+                                </Group>
+                            </Grid.Col>
+                        </Grid>
+                    </AppShell.Header>
+
+                    <AppShell.Navbar>
+                        <AppShell.Section>
+                            <Container fluid bg={colorScheme === 'dark' ? theme.colors.cyan[9] : theme.colors.blue[1]}>
+                                <Text fz="sm">
+                                    Filter existing proposals by:
+                                </Text>
+                                <TextInput label="Title"
+                                           value={proposalTitleFilter}
+                                           onChange={(e: { target: { value: string; }; }) =>
+                                               setProposalTitleFilter(e.target.value)}
+                                />
+                                <TextInput label="Investigator name"
+                                           value={investigatorNameFilter}
+                                           onChange={(e: { target: { value: string; }; }) =>
+                                               setInvestigatorNameFilter(e.target.value)}
+                                           pb={"md"}
+                                />
+                            </Container>
+                        </AppShell.Section>
+                        <AppShell.Section component={ScrollArea}>
+                            <ProposalListWrapper
+                                proposalTitle={proposalTitleFilter}
+                                investigatorName={investigatorNameFilter}
+                                auth={proposalContext.authenticated}
+                            />
+                        </AppShell.Section>
+                    </AppShell.Navbar>
+                    <AppShell.Main pr={"sm"}>
+                        <Outlet/>
+                    </AppShell.Main>
+                </AppShell>
             </ProposalContext.Provider>
         )
     }
 
-    function PSTManager() : ReactElement {
-        const cycleContext = useContext(CycleContext);
-        return (
-            <CycleContext.Provider value={cycleContext}>
-                <ManagementAppShell />
-            </CycleContext.Provider>
-        )
+    function ProposalListWrapper(props:{proposalTitle: string, investigatorName:string, auth:boolean}) : ReactElement {
+
+        if (props.auth) {
+            return <ProposalList proposalTitle={props.proposalTitle} investigatorName={props.investigatorName} />
+        }
+        else {
+            return <></>
+        }
     }
-
-
     /**
      * html to show in the main page when "proposals for username" is selected.
      * @return {ReactElement} the html to display when
@@ -252,6 +375,17 @@ function App2(): ReactElement {
                 />
             </Container>
         );
+    }
+
+    function PSTManager() : ReactElement {
+        const navigate = useNavigate();
+        return (
+            <ActionIcon
+                onClick={(e: SyntheticEvent) => {e.preventDefault(); navigate("/")}}
+            >
+                <IconLicense />
+            </ActionIcon>
+        )
     }
 }
